Virtual enviroment is ready
A virtual enviroment was fully set up with `Python 3.7.9` and `pip 20.3.3` with the following modules:
* ttkthemes
* win10toast

All of that was done so anyone around doesn't have to wonder why is this not running or what's missing.

Updated `readme.md` so you dont have to wonder what to execute and how.

Moved unnecesary .py files into `feature_testing`